+++ TwoPage/src/main/java/com/microsoft/device/display/samples/twopage/fragments/FirstPageFragment.kt
@@ -1,19 +1,19 @@
-package com.microsoft.device.display.samples.twopage
+package com.microsoft.device.display.samples.twopage.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.microsoft.device.display.samples.twopage.R
 
-class SecondPageFragment : Fragment() {
-
+class FirstPageFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_second_page, container, false)
+        return inflater.inflate(R.layout.fragment_first_page, container, false)
     }
 
 }
