+++ TwoPage/src/main/java/com/microsoft/device/display/samples/twopage/fragments/FourthPageFragment.kt
@@ -1,10 +1,11 @@
-package com.microsoft.device.display.samples.twopage
+package com.microsoft.device.display.samples.twopage.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.microsoft.device.display.samples.twopage.R
 
 class FourthPageFragment : Fragment() {
     override fun onCreateView(
